requirements.txt
feedparser==6.0.10
PyYAML==6.0.1
requests==2.32.3
